Внес правки по комментариям ревьювера: вынес в отдельные файлы хэндлеры, адаптеры, начал парсить json через gson, а не через свой метод.
