+++ src/Main.java
@@ -4,7 +4,9 @@ public class Main {
 
     public static void main(String[] args) throws IOException {
 
-        HttpTaskServer httpTaskServer = new HttpTaskServer();
+        String path = "test.csv";
+        FileBackedTaskManager fileBackedTaskManager = new FileBackedTaskManager(path);
+        HttpTaskServer httpTaskServer = new HttpTaskServer(fileBackedTaskManager);
         httpTaskServer.startServer();
     }
 }
